spliting the creation of the boarder cell into a seperate file
