+++ src/view/gameBoard/GameBoard.tsx
@@ -4,8 +4,7 @@ import {
   COLUMN_BOARD_SIZE,
   ROW_BOARD_SIZE,
 } from "../..//logic/Board/board";
-import { Cell } from "./Cell";
-
+import { RowHeadersAndCells, ColumnHeader } from "./BoardCells";
 interface GameBoardProps {
   board: Board;
 }
@@ -16,40 +15,6 @@ const boardStyle = {
   gridTemplateRows: `repeat(${ROW_BOARD_SIZE}, 30px)`,
 };
 
-const cellStyle = {
-  display: "flex",
-  alignItems: "center",
-  justifyContent: "center",
-  color: "rgb(119,119,119)",
-  fontSize: "10px",
-};
-
-const RowHeaders = (
-  colIndex: number,
-  rowIndex: number,
-  cell: number,
-  cellId: string
-) => {
-  return (
-    <>
-      {colIndex === 0 ? (
-        <>
-          <div
-            style={{
-              ...cellStyle,
-            }}
-            key={cellId}
-          >
-            R{rowIndex}
-          </div>
-        </>
-      ) : (
-        <Cell entity={cell} key={cellId} />
-      )}
-    </>
-  );
-};
-
 export const GameBoard: React.FC<GameBoardProps> = ({ board }) => {
   const columnHeader = Array.from(Array(COLUMN_BOARD_SIZE).keys());
   const newBoard: number[] | Board = [columnHeader, ...board];
@@ -58,19 +23,9 @@ export const GameBoard: React.FC<GameBoardProps> = ({ board }) => {
       const cellId = `cell-${rowIndex}-${colIndex}`;
       return (
         <>
-          {rowIndex === 0 ? (
-            <div
-              style={{
-                ...cellStyle,
-                justifySelf: "right",
-              }}
-              key={cellId}
-            >
-              C{colIndex + 1}
-            </div>
-          ) : (
-            RowHeaders(colIndex, rowIndex, cell, cellId)
-          )}
+          {rowIndex === 0
+            ? ColumnHeader(colIndex, cellId)
+            : RowHeadersAndCells(colIndex, rowIndex, cell, cellId)}
         </>
       );
     })
